+++ src/NeatMapper/Mappers/Async/AsyncIdentityMapper.cs
@@ -1,0 +1,174 @@
+﻿using System;
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace NeatMapper {
+	/// <summary>
+	/// Singleton <see cref="IAsyncMapper"/> which returns the provided source element (for both new and merge maps).
+	/// Supports only the same source/destination types. Can be used to merge collections of elements of the same type.
+	/// </summary>
+	public sealed class AsyncIdentityMapper : IAsyncMapper, IAsyncMapperCanMap, IAsyncMapperFactory {
+		/// <summary>
+		/// Singleton instance of the mapper.
+		/// </summary>
+		public static readonly IAsyncMapper Instance = new AsyncIdentityMapper();
+
+
+		internal AsyncIdentityMapper() { }
+
+
+		#region IAsyncMapper methods
+		public Task<
+#if NETCOREAPP3_1_OR_GREATER || NETSTANDARD2_1_OR_GREATER || NET5_0_OR_GREATER
+			object?
+#else
+			object
+#endif
+			> MapAsync(
+#if NETCOREAPP3_1_OR_GREATER || NETSTANDARD2_1_OR_GREATER || NET5_0_OR_GREATER
+			object?
+#else
+			object
+#endif
+			source,
+			Type sourceType,
+			Type destinationType,
+#if NETCOREAPP3_1_OR_GREATER || NETSTANDARD2_1_OR_GREATER || NET5_0_OR_GREATER
+			MappingOptions?
+#else
+			MappingOptions
+#endif
+			mappingOptions = null,
+			CancellationToken cancellationToken = default) {
+
+			if (!CanMap(sourceType, destinationType))
+				throw new MapNotFoundException((sourceType, destinationType));
+
+			TypeUtils.CheckObjectType(source, sourceType, nameof(source));
+
+			return Task.FromResult(source);
+		}
+
+		public Task<
+#if NETCOREAPP3_1_OR_GREATER || NETSTANDARD2_1_OR_GREATER || NET5_0_OR_GREATER
+			object?
+#else
+			object
+#endif
+			> MapAsync(
+#if NETCOREAPP3_1_OR_GREATER || NETSTANDARD2_1_OR_GREATER || NET5_0_OR_GREATER
+			object?
+#else
+			object
+#endif
+			source,
+			Type sourceType,
+#if NETCOREAPP3_1_OR_GREATER || NETSTANDARD2_1_OR_GREATER || NET5_0_OR_GREATER
+			object?
+#else
+			object
+#endif
+			destination,
+			Type destinationType,
+#if NETCOREAPP3_1_OR_GREATER || NETSTANDARD2_1_OR_GREATER || NET5_0_OR_GREATER
+			MappingOptions?
+#else
+			MappingOptions
+#endif
+			mappingOptions = null,
+			CancellationToken cancellationToken = default) {
+
+			if (!CanMap(sourceType, destinationType))
+				throw new MapNotFoundException((sourceType, destinationType));
+
+			TypeUtils.CheckObjectType(source, sourceType, nameof(source));
+			TypeUtils.CheckObjectType(destination, destinationType, nameof(destination));
+
+			return Task.FromResult(source);
+		}
+		#endregion
+
+		#region IAsyncMapperCanMap methods
+		public Task<bool> CanMapAsyncNew(
+			Type sourceType,
+			Type destinationType,
+#if NETCOREAPP3_1_OR_GREATER || NETSTANDARD2_1_OR_GREATER || NET5_0_OR_GREATER
+			MappingOptions?
+#else
+			MappingOptions
+#endif
+			mappingOptions = null,
+			CancellationToken cancellationToken = default) {
+
+			return Task.FromResult(CanMap(sourceType, destinationType));
+		}
+
+		public Task<bool> CanMapAsyncMerge(
+			Type sourceType,
+			Type destinationType,
+#if NETCOREAPP3_1_OR_GREATER || NETSTANDARD2_1_OR_GREATER || NET5_0_OR_GREATER
+			MappingOptions?
+#else
+			MappingOptions
+#endif
+			mappingOptions = null,
+			CancellationToken cancellationToken = default) {
+
+			return Task.FromResult(CanMap(sourceType, destinationType));
+		}
+		#endregion
+
+		#region IAsyncMapperFactory methods
+		public IAsyncNewMapFactory MapAsyncNewFactory(
+			Type sourceType,
+			Type destinationType,
+#if NETCOREAPP3_1_OR_GREATER || NETSTANDARD2_1_OR_GREATER || NET5_0_OR_GREATER
+			MappingOptions?
+#else
+			MappingOptions
+#endif
+			mappingOptions = null) {
+
+			if (!CanMap(sourceType, destinationType))
+				throw new MapNotFoundException((sourceType, destinationType));
+
+			return new DefaultAsyncNewMapFactory(sourceType, destinationType, (source, cancellationToken) => {
+				TypeUtils.CheckObjectType(source, sourceType, nameof(source));
+
+				return Task.FromResult(source);
+			});
+		}
+
+		public IAsyncMergeMapFactory MapAsyncMergeFactory(
+			Type sourceType,
+			Type destinationType,
+#if NETCOREAPP3_1_OR_GREATER || NETSTANDARD2_1_OR_GREATER || NET5_0_OR_GREATER
+			MappingOptions?
+#else
+			MappingOptions
+#endif
+			mappingOptions = null) {
+
+			if (!CanMap(sourceType, destinationType))
+				throw new MapNotFoundException((sourceType, destinationType));
+
+			return new DefaultAsyncMergeMapFactory(sourceType, destinationType, (source, destination, cancellationToken) => {
+				TypeUtils.CheckObjectType(source, sourceType, nameof(source));
+				TypeUtils.CheckObjectType(destination, destinationType, nameof(destination));
+
+				return Task.FromResult(source);
+			});
+		}
+		#endregion
+
+
+		private bool CanMap(Type sourceType, Type destinationType) {
+			if (sourceType == null)
+				throw new ArgumentNullException(nameof(sourceType));
+			if (destinationType == null)
+				throw new ArgumentNullException(nameof(destinationType));
+
+			return sourceType == destinationType;
+		}
+	}
+}