+++ tests/NeatMapper.Tests/Mapping/Async/AsyncMergeMapperTests.cs
@@ -866,7 +866,7 @@ namespace NeatMapper.Tests.Mapping.Async {
 				MappingOptionsUtils.mergeOptions = null;
 
 				var opts = new TestOptions();
-				var merge = new MergeCollectionsMappingOptions(false, (s, d, _) => false);
+				var merge = new MergeCollectionsMappingOptions(false, EmptyMatcher.Instance);
 				await _mapper.MapAsync<string>(2f, new object[]{ opts, merge });
 
 				Assert.AreSame(opts, MappingOptionsUtils.options);