+++ tests/NeatMapper.Tests/Mapping/Core/NewCollectionMapperTests.cs
@@ -71,7 +71,7 @@ namespace NeatMapper.Tests.Mapping {
 				MappingOptionsUtils.options = null;
 				MappingOptionsUtils.mergeOptions = null;
 				var opts = new TestOptions();
-				var merge = new MergeCollectionsMappingOptions(false, (s, d, _) => false);
+				var merge = new MergeCollectionsMappingOptions(false, EmptyMatcher.Instance);
 				var strings = _mapper.Map<LinkedList<string>>(new[] { 2, -3, 0 }, new object[] { opts, merge });
 
 				Assert.IsNotNull(strings);
@@ -287,7 +287,7 @@ namespace NeatMapper.Tests.Mapping {
 				MappingOptionsUtils.mergeOptions = null;
 
 				var opts = new TestOptions();
-				var merge = new MergeCollectionsMappingOptions(false, (s, d, _) => false);
+				var merge = new MergeCollectionsMappingOptions(false, EmptyMatcher.Instance);
 				_mapper.Map<IList<IEnumerable<string>>>(new[] {
 					new[]{ 2, -3, 0 },
 					new[]{ 1, 2 }